+++ src/main/kotlin/lt/skafis/bankas/config/CorsConfigDev.kt
@@ -2,20 +2,22 @@ package lt.skafis.bankas.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Profile
 import org.springframework.web.servlet.config.annotation.CorsRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
 @Configuration
-class CorsConfig {
+@Profile("dev")
+class CorsConfigDev {
     @Bean
     fun corsConfigurer(): WebMvcConfigurer {
         return object : WebMvcConfigurer {
             override fun addCorsMappings(registry: CorsRegistry) {
                 registry.addMapping("/**")
-                    .allowedOrigins("https://bankas.skafis.lt", "http://localhost:5173")
+                    .allowedOrigins("*")
+                    .allowedOrigins("*")
                     .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH")
                     .allowedHeaders("*")
-                    .allowCredentials(true)
                     .maxAge(3600)
             }
         }