+++ src/main/kotlin/lt/skafis/bankas/config/Logged.kt
@@ -1,0 +1,5 @@
+package lt.skafis.bankas.config
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Logged