+++ src/main/kotlin/lt/skafis/bankas/dto/UserBioDto.kt
@@ -1,4 +1,4 @@
-package lt.skafis.bankas.dtoOld
+package lt.skafis.bankas.dto
 
 data class UserBioDto(
     val bio: String = ""