+++ src/main/kotlin/lt/skafis/bankas/dtoOld/ProblemsForAuthor.kt
@@ -1,6 +1,0 @@
-package lt.skafis.bankas.dtoOld
-
-data class ProblemsForAuthor(
-    val underReviewProblems: List<UnderReviewProblemDisplayViewDto>,
-    val approvedProblems: List<ProblemDisplayViewDto>
-)