+++ src/main/kotlin/lt/skafis/bankas/model/Problem.kt
@@ -1,5 +1,7 @@
 package lt.skafis.bankas.model
 
+import com.google.cloud.firestore.annotation.PropertyName
+
 data class Problem (
     override var id: String = "",
     val skfCode: String = "",
@@ -9,5 +11,7 @@ data class Problem (
     val answerImagePath: String = "",
     val categoryId: String = "",
     val sourceId: String = "",
+
+    @get:PropertyName("isApproved")
     val isApproved: Boolean = false,
 ): Identifiable