+++ src/main/kotlin/lt/skafis/bankas/service/SortService.kt
@@ -1,0 +1,12 @@
+package lt.skafis.bankas.service
+
+import lt.skafis.bankas.dto.ProblemDisplayViewDto
+import lt.skafis.bankas.model.Problem
+
+interface SortService {
+    fun getMySortedProblems(): List<ProblemDisplayViewDto>
+    fun getMyUnsortedProblems(): List<ProblemDisplayViewDto>
+    fun sortProblem(problemId: String, categoryId: String): Problem
+    fun getNotMySortedProblems(): List<ProblemDisplayViewDto>
+    fun getNotMyUnsortedProblems(): List<ProblemDisplayViewDto>
+}