+++ src/main/kotlin/lt/skafis/bankas/service/implementations/ProblemMetaServiceImp.kt
@@ -2,30 +2,26 @@ package lt.skafis.bankas.service.implementations
 
 import lt.skafis.bankas.repository.MetaRepository
 import lt.skafis.bankas.service.ProblemMetaService
-import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import org.webjars.NotFoundException
 
 @Service
-class ProblemMetaServiceImp(
-    private val firestoreMetaRepository: MetaRepository
-) : ProblemMetaService {
+class ProblemMetaServiceImp: ProblemMetaService {
 
-    private val log = LoggerFactory.getLogger(javaClass)
+    @Autowired
+    private lateinit var firestoreMetaRepository: MetaRepository
 
     override fun getIncrementedLastUsedSkfCode(): String {
-        log.info("Fetching last used SKF code")
         val problemMeta = firestoreMetaRepository.getProblemMeta() ?: throw NotFoundException("Problem meta with last used SKF code not found")
         val lastUsedCode = problemMeta.lastUsedCode
         if (lastUsedCode.isBlank()) {
             throw NotFoundException("Last used SKF code in problemMeta empty or not found")
         }
-        log.info("Last used SKF code fetched successfully")
         return incrementSkf(lastUsedCode)
     }
 
     override fun incrementLastUsedSkfCode() {
-        log.info("Incrementing last used SKF code")
         val problemMeta = firestoreMetaRepository.getProblemMeta() ?: throw NotFoundException("Problem meta with last used SKF code not found")
         val lastUsedCode = problemMeta.lastUsedCode
         if (lastUsedCode.isBlank()) {
@@ -36,11 +32,10 @@ class ProblemMetaServiceImp(
         if (!success) {
             throw NotFoundException("Failed to increment last used SKF code")
         }
-        log.info("Last used SKF code incremented successfully")
     }
 
     override fun initializeLastUsedSkfCode() {
-        TODO("Not yet implemented")
+        firestoreMetaRepository.init()
     }
 
     private fun incrementSkf(skfCode: String): String {