+++ CHANGELOG.md
@@ -12,10 +12,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
-## [1.0.0] - 2024-06-22
+## [1.0.0] - 2024-07-12
 
 ### Added
-- Initial unstable release
+- Super admin access
+
+### Changed
+- Grouped by sources. Usable but unstable. Beta version
+
+## [0.0.1] - 2024-07-03
+
+### Added
+- Initial unstable category grouped release
 
 [unreleased]: https://github.com/naglissul/bankas-skafis-api/compare/v1.0.0...HEAD
-[1.0.0]: https://github.com/naglissul/bankas-skafis-api/releases/tag/v1.0.0
+[1.0.0]: https://github.com/naglissul/bankas-skafis-api/compare/v1.0.0...v0.0.1
+[0.0.1]: https://github.com/naglissul/bankas-skafis-api/releases/tag/v0.0.1